RUN-2505 : Functional tests for Key Value Data log filter
